django section/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
cramjam==2.11.0
distro==1.9.0
Django==4.2.24
django-extensions==4.1
djangorestframework==3.16.1
exceptiongroup==1.3.0
fastparquet==2024.11.0
fsspec==2025.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.11.0
numpy==2.0.2
openai==1.107.3
packaging==25.0
pandas==2.3.2
pgvector==0.4.1
psycopg2-binary==2.9.10
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
faiss-cpu
gdown